refactor: extract functions from review page to meet ~400 LoC target
Extracted entity and data loading functions into separate modules:
- entity-utils.ts: getLookupTables, getProposedEntities, calculateUnknownEntities
- data-loaders.ts: getQueueItem, getTaxonomyData, getCurrentPrompts

Result:
- page.tsx: 326 lines (was 503)
- entity-utils.ts: 95 lines
- data-loaders.ts: 93 lines

Total lines reduced while improving code organization and readability.

Files changed:
- admin-next/src/app/(dashboard)/review/[id]/page.tsx
- admin-next/src/app/(dashboard)/review/[id]/entity-utils.ts (new)
- admin-next/src/app/(dashboard)/review/[id]/data-loaders.ts (new)
