+++ admin-next/src/app/(dashboard)/review/[id]/page.tsx
@@ -1,6 +1,5 @@
 import Link from 'next/link';
 import { MarkdownRenderer } from '@/components/ui/markdown-renderer';
-import { createServiceRoleClient } from '@/lib/supabase/server';
 import { formatDateTime, getStatusColorByCode, getStatusName } from '@/lib/utils';
 import { notFound } from 'next/navigation';
 import { ReviewActions } from './actions';
@@ -9,145 +8,9 @@ import { EvaluationPanel } from './evaluation-panel';
 import { UnknownEntitiesPanel } from './unknown-entities';
 import { PipelineTimeline } from './pipeline-timeline';
 import { TagDisplay } from '@/components/tags';
-import type {
-  TaxonomyConfig,
-  TaxonomyData,
-  TaxonomyItem,
-  ValidationLookups,
-} from '@/components/tags';
-import type { QueueItem } from '@bfsi/types';
-
-// Extended QueueItem with current_run_id for this page
-interface QueueItemWithRun extends QueueItem {
-  current_run_id?: string | null;
-}
-
-interface LookupTables {
-  regulators: Set<string>;
-  standardSetters: Set<string>;
-  organizations: Set<string>;
-  vendors: Set<string>;
-}
-
-async function getQueueItem(id: string) {
-  const supabase = createServiceRoleClient();
-
-  // Try ingestion_queue first
-  const { data, error } = await supabase.from('ingestion_queue').select('*').eq('id', id).single();
-
-  if (!error && data) {
-    return data as QueueItemWithRun;
-  }
-
-  // If not found, try kb_publication (for published items)
-  const { data: pubData, error: pubError } = await supabase
-    .from('kb_publication')
-    .select(
-      'id, source_url, title, summary_short, summary_medium, summary_long, source_name, date_published, date_added, thumbnail',
-    )
-    .eq('id', id)
-    .single();
-
-  if (pubError || !pubData) {
-    return null;
-  }
-
-  // Transform kb_publication to QueueItem format
-  return {
-    id: pubData.id,
-    url: pubData.source_url,
-    status_code: 400,
-    discovered_at: pubData.date_added || '',
-    payload: {
-      title: pubData.title,
-      source_name: pubData.source_name,
-      date_published: pubData.date_published,
-      thumbnail_url: pubData.thumbnail,
-      summary: {
-        short: pubData.summary_short,
-        medium: pubData.summary_medium,
-        long: pubData.summary_long,
-      },
-    },
-  } as QueueItemWithRun;
-}
-
-async function getLookupTables(): Promise<LookupTables> {
-  const supabase = createServiceRoleClient();
-
-  const [regulatorsRes, standardSettersRes, orgsRes, vendorsRes] = await Promise.all([
-    supabase.from('regulator').select('slug'),
-    supabase.from('standard_setter').select('slug'),
-    supabase.from('bfsi_organization').select('slug'),
-    supabase.from('ag_vendor').select('slug'),
-  ]);
-
-  return {
-    regulators: new Set((regulatorsRes.data || []).map((r) => r.slug)),
-    standardSetters: new Set((standardSettersRes.data || []).map((s) => s.slug)),
-    organizations: new Set((orgsRes.data || []).map((o) => o.slug)),
-    vendors: new Set((vendorsRes.data || []).map((v) => v.slug)),
-  };
-}
-
-async function getProposedEntities(sourceQueueId: string) {
-  const supabase = createServiceRoleClient();
-
-  const { data } = await supabase
-    .from('proposed_entity')
-    .select('entity_type, name, slug')
-    .eq('source_queue_id', sourceQueueId)
-    .eq('status', 'pending');
-
-  return data || [];
-}
-
-async function getTaxonomyData() {
-  const supabase = createServiceRoleClient();
-
-  const { data: configData } = await supabase
-    .from('taxonomy_config')
-    .select(
-      'slug, display_name, display_order, behavior_type, source_table, payload_field, color, score_parent_slug, score_threshold',
-    )
-    .eq('is_active', true)
-    .order('display_order');
-
-  const taxonomyConfig = (configData || []) as TaxonomyConfig[];
-
-  const taxonomyData: TaxonomyData = {};
-  const sourceTables = taxonomyConfig
-    .filter((c) => c.source_table && c.behavior_type !== 'scoring')
-    .map((c) => ({ slug: c.slug, table: c.source_table! }));
-
-  const tableResults = await Promise.all(
-    sourceTables.map(({ slug, table }) =>
-      supabase
-        .from(table)
-        .select('code, name')
-        .order('name')
-        .then((res) => ({ slug, data: res.data || [] })),
-    ),
-  );
-
-  for (const { slug, data } of tableResults) {
-    taxonomyData[slug] = data as TaxonomyItem[];
-  }
-
-  return { taxonomyConfig, taxonomyData };
-}
-
-async function getCurrentPrompts() {
-  const supabase = createServiceRoleClient();
-
-  const { data } = await supabase
-    .from('prompt_version')
-    .select('id, version, agent_name')
-    .eq('is_current', true)
-    .in('agent_name', ['summarizer', 'tagger', 'thumbnail-generator']);
-
-  return (data || []) as { id: string; version: string; agent_name: string }[];
-}
+import type { ValidationLookups } from '@/components/tags';
+import { getLookupTables, getProposedEntities, calculateUnknownEntities } from './entity-utils';
+import { getQueueItem, getTaxonomyData, getCurrentPrompts } from './data-loaders';
 
 export default async function ReviewDetailPage({
   params,
@@ -183,47 +46,7 @@ export default async function ReviewDetailPage({
   };
 
   // Calculate unknown entities
-  const unknownEntities: {
-    entityType: 'regulator' | 'standard_setter' | 'bfsi_organization' | 'ag_vendor';
-    name: string;
-    label: string;
-  }[] = [];
-
-  // Build set of already-proposed entities to filter out
-  const proposedSet = new Set(
-    proposedEntities.map((p) => `${p.entity_type}:${p.name.toLowerCase()}`),
-  );
-
-  const isAlreadyProposed = (entityType: string, name: string) => {
-    return proposedSet.has(`${entityType}:${name.toLowerCase()}`);
-  };
-
-  for (const name of (payload.organization_names as string[]) || []) {
-    if (
-      !lookups.organizations.has(name.toLowerCase()) &&
-      !isAlreadyProposed('bfsi_organization', name)
-    ) {
-      unknownEntities.push({ entityType: 'bfsi_organization', name, label: 'Organization' });
-    }
-  }
-  for (const name of (payload.vendor_names as string[]) || []) {
-    if (!lookups.vendors.has(name.toLowerCase()) && !isAlreadyProposed('ag_vendor', name)) {
-      unknownEntities.push({ entityType: 'ag_vendor', name, label: 'Vendor' });
-    }
-  }
-  for (const code of (payload.regulator_codes as string[]) || []) {
-    if (!lookups.regulators.has(code.toLowerCase()) && !isAlreadyProposed('regulator', code)) {
-      unknownEntities.push({ entityType: 'regulator', name: code, label: 'Regulator' });
-    }
-  }
-  for (const code of (payload.standard_setter_codes as string[]) || []) {
-    if (
-      !lookups.standardSetters.has(code.toLowerCase()) &&
-      !isAlreadyProposed('standard_setter', code)
-    ) {
-      unknownEntities.push({ entityType: 'standard_setter', name: code, label: 'Standard Setter' });
-    }
-  }
+  const unknownEntities = calculateUnknownEntities(payload, lookups, proposedEntities);
 
   return (
     <div className="space-y-6">
